# final-exam-wads/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Simple Blog API — WADS Final Exam",
    "version": "1.0.0",
    "description": "REST API for blog posts, authentication (Better Auth), and AI summarize."
  },
  "servers": [{ "url": "http://localhost:3000", "description": "Local dev" }],
  "paths": {
    "/api/posts": {
      "get": {
        "summary": "List posts for authenticated user",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "List of posts" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a post",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "published": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "summary": "Get post by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Post" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update post",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "summary": "Delete post",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/api/ai/summarize": {
      "post": {
        "summary": "Summarize blog content with local Ollama",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string", "minLength": 10 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Summary generated" },
          "401": { "description": "Unauthorized" },
          "504": { "description": "AI timeout" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    }
  }
}